+++ src/sandbox.ml
@@ -56,11 +56,24 @@ module Package = struct
   ;;
 end
 
+(** If [Workspace.workspaceFolders()] returns a list with a single element,
+    returns it; otherwise, returns [None]. *)
+let workspace_root () =
+  match Workspace.workspaceFolders () with
+  | [] -> None
+  | [ workspace_folder ] ->
+    Some (workspace_folder |> WorkspaceFolder.uri |> Uri.path |> Path.of_string)
+  | _ ->
+    (* We don't support multiple workspace roots at the moment *)
+    None
+;;
+
 type t =
   | Opam of Opam.t * Opam.Switch.t
   | Esy of Esy.t * Esy.Manifest.t
   | Global
   | Custom of string
+  | Dune of Dune.t
 
 let equal t1 t2 =
   match t1, t2 with
@@ -68,6 +81,7 @@ let equal t1 t2 =
   | Esy (e1, p1), Esy (e2, p2) -> Esy.Manifest.equal p1 p2 && Esy.equal e1 e2
   | Opam (o1, s1), Opam (o2, s2) -> Opam.Switch.equal s1 s2 && Opam.equal o1 o2
   | Custom s1, Custom s2 -> String.equal s1 s2
+  | Dune d1, Dune d2 -> Dune.equal d1 d2
   | _, _ -> false
 ;;
 
@@ -76,6 +90,7 @@ let to_string = function
   | Opam (_, switch) -> Printf.sprintf "opam(%s)" (Opam.Switch.name switch)
   | Global -> "global"
   | Custom _ -> "custom"
+  | Dune _ -> "dune"
 ;;
 
 let to_pretty_string t =
@@ -91,6 +106,7 @@ let to_pretty_string t =
     print_opam project_name
   | Global -> "Global OCaml"
   | Custom _ -> "Custom OCaml"
+  | Dune _ -> "Dune Package Manager"
 ;;
 
 module Kind = struct
@@ -99,12 +115,14 @@ module Kind = struct
     | Esy
     | Global
     | Custom
+    | Dune
 
   let of_string = function
     | "opam" -> Some Opam
     | "esy" -> Some Esy
     | "global" -> Some Global
     | "custom" -> Some Custom
+    | "dune" -> Some Dune
     | _ -> None
   ;;
 
@@ -121,6 +139,7 @@ module Kind = struct
     | Esy -> "esy"
     | Global -> "global"
     | Custom -> "custom"
+    | Dune -> "dune"
   ;;
 
   let to_json s = Jsonoo.Encode.string (to_string s)
@@ -132,12 +151,14 @@ module Setting = struct
     | Esy of Esy.Manifest.t
     | Global
     | Custom of string
+    | Dune of Path.t
 
   let kind : t -> Kind.t = function
     | Opam _ -> Opam
     | Esy _ -> Esy
     | Global -> Global
     | Custom _ -> Custom
+    | Dune _ -> Dune
   ;;
 
   let of_json json =
@@ -161,6 +182,9 @@ module Setting = struct
     | Custom ->
       let template = Jsonoo.Decode.field "template" decode_vars json in
       Custom template
+    | Dune ->
+      let root = Jsonoo.Decode.field "root" decode_vars json |> Path.of_string in
+      Dune root
   ;;
 
   let to_json (t : t) =
@@ -174,6 +198,7 @@ module Setting = struct
         [ kind; "root", encode_vars @@ (manifest |> Esy.Manifest.path |> Path.to_string) ]
     | Opam switch -> object_ [ kind; "switch", encode_vars @@ Opam.Switch.name switch ]
     | Custom template -> object_ [ kind; "template", string template ]
+    | Dune dune -> object_ [ kind; "root", encode_vars @@ Path.to_string dune ]
   ;;
 
   let t = Settings.create_setting ~scope:Workspace ~key:"sandbox" ~of_json ~to_json
@@ -182,10 +207,11 @@ end
 type available_sandboxes =
   { opam : Opam.t option Promise.t
   ; esy : Esy.t option Promise.t
+  ; dune : Dune.t option Promise.t
   }
 
 let available_sandboxes () : available_sandboxes =
-  { opam = Opam.make (); esy = Esy.make () }
+  { dune = Dune.make (workspace_root ()) (); opam = Opam.make (); esy = Esy.make () }
 ;;
 
 let of_settings () : t option Promise.t =
@@ -196,6 +222,7 @@ let of_settings () : t option Promise.t =
       match kind with
       | `Esy -> "esy"
       | `Opam -> "opam"
+      | `Dune -> "dune"
     in
     show_message
       `Warn
@@ -231,18 +258,14 @@ let of_settings () : t option Promise.t =
          None))
   | Some Global -> Promise.return (Some Global)
   | Some (Custom template) -> Promise.return (Some (Custom template))
-;;
-
-(** If [Workspace.workspaceFolders()] returns a list with a single element,
-    returns it; otherwise, returns [None]. *)
-let workspace_root () =
-  match Workspace.workspaceFolders () with
-  | [] -> None
-  | [ workspace_folder ] ->
-    Some (workspace_folder |> WorkspaceFolder.uri |> Uri.path |> Path.of_string)
-  | _ ->
-    (* We don't support multiple workspace roots at the moment *)
-    None
+  | Some (Dune _dune) ->
+    let open Promise.Syntax in
+    let* dune = available.dune in
+    (match dune with
+     | None ->
+       not_available `Dune;
+       Promise.return None
+     | Some dune -> Promise.return (Some (Dune dune)))
 ;;
 
 let detect_esy_sandbox ~project_root esy () =
@@ -282,6 +305,16 @@ let detect_opam_sandbox ~project_root opam () =
   Opam (opam, switch)
 ;;
 
+let detect_dune_pkg ~project_root _dune () =
+  let open Promise.Syntax in
+  Dune.make (Some project_root) ()
+  >>= function
+  | Some dune ->
+    Dune.is_project_locked dune
+    >>| fun project_locked -> if project_locked then Some (Dune dune) else None
+  | None -> Promise.return None
+;;
+
 let detect () =
   let open Promise.Option.Syntax in
   let* project_root = workspace_root () |> Promise.return in
@@ -291,6 +324,7 @@ let detect () =
     [ detect_opam_local_switch ~project_root available.opam
     ; detect_esy_sandbox ~project_root available.esy
     ; detect_opam_sandbox ~project_root available.opam
+    ; detect_dune_pkg ~project_root available.dune
     ]
 ;;
 
@@ -308,6 +342,7 @@ let save_to_settings sandbox =
     | Opam (_, switch) -> Setting.Opam switch
     | Global -> Setting.Global
     | Custom template -> Setting.Custom template
+    | Dune dune -> Setting.Dune (Dune.root dune)
   in
   Settings.set ~section:"ocaml" Setting.t (to_setting sandbox)
 ;;
@@ -335,7 +370,7 @@ module Candidate = struct
            then Some (switch_kind_s ^ " | Currently active switch in project root")
            else Some switch_kind_s
          | Esy (_, _) -> Some "Esy"
-         | Global | Custom _ -> None)
+         | Global | Custom _ | Dune _ -> None)
     in
     match sandbox with
     | Opam (_, Named name) -> create ~label:name ?description ()
@@ -360,19 +395,25 @@ module Candidate = struct
         ~label:"Custom"
         ~detail:"Custom sandbox using a command template"
         ()
+    | Dune dune ->
+      let project_path = Path.to_string (Dune.root dune) in
+      create ?description ~label:"Dune Package Manager" ~detail:project_path ()
   ;;
 
   let ok sandbox = { sandbox; status = Ok () }
 end
 
-let select_sandbox (choices : Candidate.t list) =
+let select_sandbox (choices : Candidate.t list) t =
   let placeHolder = "Which package manager would you like to manage the sandbox?" in
   let open Promise.Syntax in
   let* current_switch =
-    let open Promise.Option.Syntax in
-    let* opam = Opam.make () in
-    let* cwd = workspace_root () |> Promise.return in
-    Opam.switch_show ~cwd opam
+    match t with
+    | Dune _ -> Promise.return None
+    | _ ->
+      let open Promise.Option.Syntax in
+      let* opam = Opam.make () in
+      let* cwd = workspace_root () |> Promise.return in
+      Opam.switch_show ~cwd opam
   in
   let choices =
     List.map
@@ -443,18 +484,32 @@ let sandbox_candidates ~workspace_folders =
     (* doesn't matter what the custom fields are set to here user will input
        custom commands in [select] *)
   in
-  let+ esy, (opam, current_switch) = Promise.all2 (esy, opam) in
-  let cs = (global :: custom :: esy) @ opam in
+  let dune =
+    let+ dune_projects =
+      workspace_folders
+      |> List.map ~f:(fun (folder : WorkspaceFolder.t) ->
+        let root = folder |> WorkspaceFolder.uri |> Uri.fsPath |> Path.of_string in
+        let+ dune = Dune.make (Some root) () in
+        match dune with
+        | Some dune -> [ dune ]
+        | None -> [])
+      |> Promise.all_list
+    in
+    List.concat dune_projects
+    |> List.map ~f:(fun dune -> { Candidate.sandbox = Dune dune; status = Ok () })
+  in
+  let+ esy, (opam, current_switch), dune = Promise.all3 (esy, opam, dune) in
+  let cs = (global :: custom :: dune) @ esy @ opam in
   Option.value_map current_switch ~default:cs ~f:(fun current_switch ->
     current_switch :: cs)
 ;;
 
-let select_sandbox () =
+let select_sandbox t =
   let open Promise.Syntax in
   let workspace_folders = Workspace.workspaceFolders () in
   let* candidates = sandbox_candidates ~workspace_folders in
   let open Promise.Option.Syntax in
-  let* candidate = select_sandbox candidates in
+  let* candidate = select_sandbox candidates t in
   match candidate with
   | { status = Ok (); sandbox = Custom _ } ->
     let validateInput ~value =
@@ -482,18 +537,30 @@ let select_sandbox () =
      | Ok () -> Promise.Option.return sandbox)
 ;;
 
-let select_sandbox_and_save () =
+let select_sandbox_and_save t =
   let open Promise.Option.Syntax in
-  let* sandbox = select_sandbox () in
+  let* sandbox = select_sandbox t in
   let open Promise.Syntax in
   let+ () = save_to_settings sandbox in
   Some sandbox
 ;;
 
-let get_command sandbox bin args : Cmd.t =
+let get_command
+      sandbox
+      bin
+      args
+      (dune_cmd_type : [> `Tool | `Command | `Exec | `Install ])
+  : Cmd.t
+  =
   match sandbox with
   | Opam (opam, switch) -> Opam.exec opam switch ~args:(bin :: args)
   | Esy (esy, manifest) -> Esy.exec esy manifest ~args:(bin :: args)
+  | Dune dune ->
+    (match dune_cmd_type with
+     | `Install -> Dune.tools ~tool:bin ~args dune `Install
+     | `Tool -> Dune.tools ~tool:bin ~args dune `Exec_
+     | `Command -> Dune.command dune ~args
+     | `Exec -> Dune.exec ~target:bin ~args dune)
   | Global -> Spawn { bin = Path.of_string bin; args }
   | Custom template ->
     let command =
@@ -520,7 +587,7 @@ let get_exec_command sandbox tools =
 ;;
 
 let ocaml_version sandbox =
-  let cmd = get_command sandbox "ocamlc" [ "--version" ] in
+  let cmd = get_command sandbox "ocamlc" [ "--version" ] `Exec in
   let open Promise.Result.Syntax in
   let* cmd = Cmd.check cmd in
   let+ output = Cmd.output cmd in
@@ -531,7 +598,7 @@ let packages t =
   let open Promise.Result.Syntax in
   match t with
   | Global -> Promise.Result.return []
-  | Custom _ -> Promise.Result.return []
+  | Custom _ | Dune _ -> Promise.Result.return []
   | Esy (esy, manifest) ->
     let+ r = Esy.packages esy manifest in
     List.map r ~f:Package.of_esy
@@ -544,7 +611,7 @@ let root_packages t =
   let open Promise.Result.Syntax in
   match t with
   | Global -> Promise.Result.return []
-  | Custom _ -> Promise.Result.return []
+  | Custom _ | Dune _ -> Promise.Result.return []
   | Esy (esy, manifest) ->
     let+ r = Esy.root_packages esy manifest in
     List.map r ~f:Package.of_esy
@@ -567,6 +634,11 @@ let uninstall_packages t packages =
     Promise.return ()
   | Custom _ ->
     show_message `Error "Uninstalling packages is not supported for Custom sandboxes";
+    Promise.return ()
+  | Dune _ ->
+    show_message
+      `Error
+      "Uninstalling packages is not yet supported for Dune Package Manager";
     Promise.return ()
   | Esy (_esy, _manifest) ->
     (* TODO: Implement Esy sandbox inspection *)
@@ -613,6 +685,11 @@ let install_packages t packages =
   | Custom _ ->
     show_message `Error "Installing packages is not supported for Custom sandboxes";
     Promise.return ()
+  | Dune _ ->
+    show_message
+      `Error
+      "Installing packages is not yet supported for Dune Package Manager";
+    Promise.return ()
   | Esy (_esy, _manifest) ->
     (* TODO: Implement Esy sandbox inspection *)
     show_message `Error "Installing packages is not supported for Esy sandboxes";
@@ -651,6 +728,9 @@ let upgrade_packages ?(packages = []) t =
     Promise.return ()
   | Custom _ ->
     show_message `Error "Upgrading packages is not supported for Custom sandboxes";
+    Promise.return ()
+  | Dune _ ->
+    show_message `Error "Upgrading packages is not yet supported for Dune Package Manager";
     Promise.return ()
   | Esy (_esy, _manifest) ->
     (* TODO: Implement Esy sandbox inspection *)